Move running script to a separate function
